requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
bokeh==3.6.3
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorcet==3.1.0
contourpy==1.3.2
cucim-cu12==25.6.0
cuda-bindings==12.9.0
cuda-python==12.9.0
cudf-cu12==25.6.0
cugraph-cu12==25.6.0
cuml-cu12==25.6.0
cupy-cuda12x==13.4.1
cuvs-cu12==25.6.1
cuxfilter-cu12==25.6.0
dask==2025.5.0
dask-cuda==25.6.0
dask-cudf-cu12==25.6.0
datashader==0.18.1
defusedxml==0.7.1
distributed==2025.5.0
distributed-ucxx-cu12==0.44.0
fastjsonschema==2.21.1
fastrlock==0.8.3
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
geopandas==1.1.1
holoviews==1.21.0
idna==3.10
imageio==2.37.0
isoduration==20.11.0
Jinja2==3.1.6
joblib==1.5.1
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
lazy_loader==0.4
libcudf-cu12==25.6.0
libcugraph-cu12==25.6.0
libcuml-cu12==25.6.0
libcuvs-cu12==25.6.1
libkvikio-cu12==25.6.0
libraft-cu12==25.6.0
librmm-cu12==25.6.0
libucx-cu12==1.18.1
libucxx-cu12==0.44.0
linkify-it-py==2.0.3
llvmlite==0.44.0
locket==1.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
msgpack==1.1.1
multidict==6.6.3
multipledispatch==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
numba==0.61.2
numba-cuda==0.11.0
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.9.86
nvtx==0.2.12
nx-cugraph-cu12==25.6.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
panel==1.7.2
param==2.2.1
partd==1.4.2
pillow==11.3.0
platformdirs==4.3.8
prometheus_client==0.22.1
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pyarrow==19.0.1
pycparser==2.22
pyct==0.5.0
Pygments==2.19.2
pylibcudf-cu12==25.6.0
pylibcugraph-cu12==25.6.0
pylibraft-cu12==25.6.0
pynvjitlink-cu12==0.7.0
pynvml==12.0.0
pyogrio==0.11.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.2
pyzmq==27.0.0
raft-dask-cu12==25.6.0
rapids-dask-dependency==25.6.0
rapids-logger==0.1.1
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rmm-cu12==25.6.0
rpds-py==0.26.0
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
Send2Trash==1.8.3
shapely==2.0.7
simpervisor==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tblib==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.6.11
tinycss2==1.4.0
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
treelite==4.4.1
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
uc-micro-py==1.0.3
ucx-py-cu12==0.44.0
ucxx-cu12==0.44.0
uri-template==1.3.0
urllib3==2.5.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xarray==2025.7.0
xyzservices==2025.4.0
yarl==1.20.1
zict==3.0.0
